task7-8/digital-signature-app/requirements.txt
cryptography
PyQt5
PySide2
pytest
numpy